+++ tests/acceptance/register-test.ts
@@ -1,5 +1,6 @@
 import { currentURL, visit } from '@ember/test-helpers';
 import setupMirage from 'ember-cli-mirage/test-support/setup-mirage';
+import { percySnapshot } from 'ember-percy';
 import { setupApplicationTest } from 'ember-qunit';
 import { module, test } from 'qunit';
 
@@ -18,6 +19,7 @@ module('Acceptance | register (sign up page)', hooks => {
         assert.dom('[data-test-sign-up-full-name]').exists();
 
         assert.dom('img[alt*="Missing translation"]').doesNotExist();
+        await percySnapshot(assert);
     });
 
     test('visiting /register?next=foo', async assert => {